# TSUS.UI/TSUS.UI/swagger.json
{
    "openapi": "3.0.1",
    "info": {
      "title": "TSUS.API",
      "version": "1.0"
    },
    "servers": [
      {
        "url": "https://localhost:7078",
        "description": "Development Server"
      }
    ],
    "paths": {
      "/Auth/Registration": {
        "post": {
          "tags": [
            "Auth"
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistrationDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistrationDto"
                }
              },
              "application/*+json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistrationDto"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Success"
            }
          }
        }
      },
      "/Auth/Login": {
        "post": {
          "tags": [
            "Auth"
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginDto"
                }
              },
              "application/*+json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginDto"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Success",
              "content": {
                "text/plain": {
                  "schema": {
                    "type": "string"
                  }
                },
                "application/json": {
                  "schema": {
                    "type": "string"
                  }
                },
                "text/json": {
                  "schema": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      },
      "/Auth/Email": {
        "post": {
          "tags": [
            "Auth"
          ],
          "parameters": [
            {
              "name": "email",
              "in": "query",
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Success"
            }
          }
        }
      },
      "/Auth/Verify": {
        "put": {
          "tags": [
            "Auth"
          ],
          "parameters": [
            {
              "name": "email",
              "in": "query",
              "schema": {
                "type": "string"
              }
            },
            {
              "name": "code",
              "in": "query",
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Success"
            }
          }
        }
      },
      "/Auth/RequestPasswordReset": {
        "get": {
          "tags": [
            "Auth"
          ],
          "parameters": [
            {
              "name": "email",
              "in": "query",
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Success",
              "content": {
                "text/plain": {
                  "schema": {
                    "type": "integer",
                    "format": "int32"
                  }
                },
                "application/json": {
                  "schema": {
                    "type": "integer",
                    "format": "int32"
                  }
                },
                "text/json": {
                  "schema": {
                    "type": "integer",
                    "format": "int32"
                  }
                }
              }
            }
          }
        }
      },
      "/Auth/ChangePassword": {
        "put": {
          "tags": [
            "Auth"
          ],
          "parameters": [
            {
              "name": "userId",
              "in": "query",
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            },
            {
              "name": "password",
              "in": "query",
              "schema": {
                "type": "string"
              }
            },
            {
              "name": "code",
              "in": "query",
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Success"
            }
          }
        }
      },
      "/Departments": {
        "get": {
          "tags": [
            "Departments"
          ],
          "parameters": [
            {
              "name": "controlFlags",
              "in": "query",
              "schema": {
                "$ref": "#/components/schemas/BaseControlFlags"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Success",
              "content": {
                "text/plain": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/DepartmentRm"
                    }
                  }
                },
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/DepartmentRm"
                    }
                  }
                },
                "text/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/DepartmentRm"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "/Departments/Add": {
        "post": {
          "tags": [
            "Departments"
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DepartmentDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DepartmentDto"
                }
              },
              "application/*+json": {
                "schema": {
                  "$ref": "#/components/schemas/DepartmentDto"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Success"
            }
          }
        }
      },
      "/Faculties/Add": {
        "post": {
          "tags": [
            "Faculties"
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FacultyDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/FacultyDto"
                }
              },
              "application/*+json": {
                "schema": {
                  "$ref": "#/components/schemas/FacultyDto"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Success"
            }
          }
        }
      }
    },
    "components": {
      "schemas": {
        "BaseControlFlags": {
          "enum": [
            0,
            1
          ],
          "type": "integer",
          "format": "int32"
        },
        "DepartmentDto": {
          "required": [
            "facultyId",
            "name"
          ],
          "type": "object",
          "properties": {
            "name": {
              "maxLength": 50,
              "minLength": 0,
              "type": "string"
            },
            "facultyId": {
              "type": "integer",
              "format": "int32"
            }
          },
          "additionalProperties": false
        },
        "DepartmentRm": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer",
              "format": "int32"
            },
            "name": {
              "type": "string",
              "nullable": true
            },
            "department": {
              "type": "string",
              "nullable": true
            }
          },
          "additionalProperties": false
        },
        "FacultyDto": {
          "required": [
            "name"
          ],
          "type": "object",
          "properties": {
            "name": {
              "maxLength": 50,
              "minLength": 0,
              "type": "string"
            }
          },
          "additionalProperties": false
        },
        "LoginDto": {
          "required": [
            "email",
            "password"
          ],
          "type": "object",
          "properties": {
            "email": {
              "minLength": 1,
              "type": "string"
            },
            "password": {
              "minLength": 1,
              "type": "string"
            }
          },
          "additionalProperties": false
        },
        "RegistrationDto": {
          "required": [
            "email",
            "password",
            "userName"
          ],
          "type": "object",
          "properties": {
            "email": {
              "maxLength": 50,
              "minLength": 0,
              "type": "string"
            },
            "userName": {
              "maxLength": 50,
              "minLength": 0,
              "type": "string"
            },
            "password": {
              "maxLength": 25,
              "minLength": 0,
              "type": "string"
            },
            "departmentId": {
              "type": "integer",
              "format": "int32"
            }
          },
          "additionalProperties": false
        }
      },
      "securitySchemes": {
        "oauth2": {
          "type": "apiKey",
          "description": "Standard Authorization Header using the Bearer scheme (\"bearer {token}\")",
          "name": "Authorization",
          "in": "header"
        }
      }
    }
  }